[REF] reward, *: combine reward concept between coupons and loyalty
Before this commit the coupon rewards were handled separately from
loyalty rewards, preventing the calculation from taking the other one
into account.

After this commit the coupon rewards and the loyalty rewards are relying
on the same common reward concept.

task-2319729
PR-63628
Enterprise PR-15375
