+++ addons/loyalty/data/loyalty_demo.xml
@@ -10,9 +10,9 @@
 
     <record id="reward_01" model="loyalty.reward">
         <field name="name">Free Pen</field>
-        <field name="reward_type">gift</field>
+        <field name="reward_type">product</field>
         <field name="point_cost">5</field>
-        <field name="gift_product_id" ref="loyalty.simple_pen" />
+        <field name="reward_product_id" ref="loyalty.simple_pen" />
     </record>
 
     <record id="reward_02" model="loyalty.reward">
