+++ addons/loyalty/models/product.py
@@ -11,7 +11,7 @@ class ProductProduct(models.Model):
     def write(self, vals):
         if 'active' in vals and not vals['active']:
             product_in_reward = self.env['loyalty.reward'].sudo().search(['|',
-                    ('gift_product_id', 'in', self.ids),
+                    ('reward_product_id', 'in', self.ids),
                     ('discount_product_id', 'in', self.ids)], limit=1)
             if product_in_reward:
                 raise ValidationError(_("The product cannot be archived because it's used in a loyalty program."))
