+++ addons/reward/__manifest__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+{
+    'name': "Reward",
+    'summary': "Rewards for coupons and loyalty programs.",
+    'description': """Introduce a reward concept that can be used in other modules.""",
+    'category': 'Sales',
+    'version': '1.0',
+    'depends': ['product'],
+    'data': [
+    ],
+    'installable': True,
+    'license': 'LGPL-3',
+}
