+++ addons/sale_reward/__manifest__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+{
+    'name': "Sale Reward",
+    'summary': "Use rewards in sales orders",
+    'description': """Integrate reward mechanism in sales orders.""",
+    'category': 'Sales/Sales',
+    'version': '1.0',
+    'depends': ['reward', 'sale'],
+    'data': [
+    ],
+    'license': 'LGPL-3',
+}
